feat: Configurable working directory for PNP processors (#55)
Add new 'pnp-build-processor-nested' - aimed for the multi-processors repositories. I.e. having a single repo with a folder for each of the fully decoupled processors.
In this case single repo is used only for logical grouping, simplifying search, etc. Processors are completely atomic, each folder could be considered it's own repo.